finished mapTerrain
finished the mapTerrrain function, Also I noticed that I had the incorrect image being used. I was using the proffesors output example rather than input. This led to there being a red line for the path which was causing problems when I was mapping the different colors. I kept finding red and being confused how it was there. I was able to highlight the red using matplotlib to figure out what was going on.